+++ Model/Api/GetCheckoutData.php
@@ -14,27 +14,30 @@ class GetCheckoutData implements GetCheckoutDataInterface
     private $readerWriter;
     private $requestFactory;
     private $moduleConfig;
-    private $jsonResultFactory;
+//    private $jsonResultFactory;
     private $scopeConfig;
     private $apiRequest;
     private $paymentsPlans;
+    private $configProvider;
     
     public function __construct(
         Config $moduleConfig,
         \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
-        \Magento\Framework\Controller\Result\JsonFactory $jsonResultFactory,
+//        \Magento\Framework\Controller\Result\JsonFactory $jsonResultFactory,
         \Magento\Framework\Webapi\Rest\Request $apiRequest,
         \Nuvei\Checkout\Model\PaymentsPlans $paymentsPlans,
         \Nuvei\Checkout\Model\ReaderWriter $readerWriter,
-        \Nuvei\Checkout\Model\Request\Factory $requestFactory
+        \Nuvei\Checkout\Model\Request\Factory $requestFactory,
+        \Nuvei\Checkout\Model\ConfigProvider $configProvider
     ) {
         $this->readerWriter         = $readerWriter;
         $this->moduleConfig         = $moduleConfig;
         $this->requestFactory       = $requestFactory;
-        $this->jsonResultFactory    = $jsonResultFactory;
+//        $this->jsonResultFactory    = $jsonResultFactory;
         $this->scopeConfig          = $scopeConfig;
         $this->apiRequest           = $apiRequest;
         $this->paymentsPlans        = $paymentsPlans;
+        $this->configProvider       = $configProvider;
     }
     
     /**
@@ -45,13 +48,11 @@ class GetCheckoutData implements GetCheckoutDataInterface
      */
     public function getData($quoteId, $neededData)
     {
-        $this->readerWriter->createLog(
-            [
+        $this->readerWriter->createLog([
             $quoteId, 
             $neededData, 
             $this->apiRequest->getBodyParams()
-            ]
-        );
+        ]);
         
         //        $result = $this->jsonResultFactory->create()
         //            ->setHttpResponseCode(\Magento\Framework\Webapi\Response::HTTP_OK);
@@ -190,6 +191,9 @@ class GetCheckoutData implements GetCheckoutDataInterface
             \Magento\Store\Model\ScopeInterface::SCOPE_STORE
         );
         
+        /**
+         * TODO - try to get configurations from ConfigProvider class as in getSimplyConnectData()!
+         */
         $sdk_data = [
             'sessionToken'      => $oo_data['sessionToken'],
             'total'             => $oo_data['amount'],
@@ -276,43 +280,44 @@ class GetCheckoutData implements GetCheckoutDataInterface
             'isTestMode'            => $isTestMode,
             'countryId'             => $this->moduleConfig->getQuoteCountryCode($quoteId),
             'isPaymentPlan'         => $isPaymentPlan,
-            'nuveiCheckoutParams'   => [
-                'env'                       => $isTestMode ? 'test' : 'prod',
-                'merchantId'                => $this->moduleConfig->getMerchantId(),
-                'merchantSiteId'            => $this->moduleConfig->getMerchantSiteId(),
-                'country'                   => $billingAddress['country'],
-                'currency'                  => $this->moduleConfig->getQuoteBaseCurrency($quoteId),
-                'amount'                    => $this->moduleConfig->getQuoteBaseTotal($quoteId),
-                'renderTo'                  => '#nuvei_checkout',
-                'useDCC'                    =>  $this->moduleConfig->getConfigValue('use_dcc'),
-                'strict'                    => false,
-                'savePM'                    => $save_pm,
-                'showUserPaymentOptions'    => ($isUserLogged && $this->moduleConfig->canShowUpos()) ? true : false,
-                'alwaysCollectCvv'          => true,
-                'fullName'                  => trim((string) $billingAddress['firstName'] . ' ' 
-                    . (string) $billingAddress['lastName']),
-                'email'                     => $billingAddress['email'],
-                'payButton'                 => $this->moduleConfig->getConfigValue('pay_btn_text'),
-                'showResponseMessage'       => false, // shows/hide the response popups
-                'locale'                    => substr($locale, 0, 2),
-                'autoOpenPM'                => (bool) $this->moduleConfig->getConfigValue('auto_expand_pms'),
-                'logLevel'                  => $this->moduleConfig->getConfigValue('checkout_log_level'),
-                'maskCvv'                   => true,
-                'i18n'                      => $this->moduleConfig->getCheckoutTransl(),
-                'blockCards'                => $blocked_cards,
-                'theme'                     => $this->moduleConfig->getConfigValue('sdk_theme', 'checkout'),
-                'apmWindowType'             => $this->moduleConfig->getConfigValue('apm_window_type', 'checkout'),
-            ],
+            'nuveiCheckoutParams'   => $this->configProvider->getConfig(),
+//            [
+//                'env'                       => $isTestMode ? 'test' : 'prod',
+//                'merchantId'                => $this->moduleConfig->getMerchantId(),
+//                'merchantSiteId'            => $this->moduleConfig->getMerchantSiteId(),
+//                'country'                   => $billingAddress['country'],
+//                'currency'                  => $this->moduleConfig->getQuoteBaseCurrency($quoteId),
+//                'amount'                    => $this->moduleConfig->getQuoteBaseTotal($quoteId),
+//                'renderTo'                  => '#nuvei_checkout',
+//                'useDCC'                    =>  $this->moduleConfig->getConfigValue('use_dcc'),
+//                'strict'                    => false,
+//                'savePM'                    => $save_pm,
+//                'showUserPaymentOptions'    => ($isUserLogged && $this->moduleConfig->canShowUpos()) ? true : false,
+//                'alwaysCollectCvv'          => true,
+//                'fullName'                  => trim((string) $billingAddress['firstName'] . ' ' 
+//                    . (string) $billingAddress['lastName']),
+//                'email'                     => $billingAddress['email'],
+//                'payButton'                 => $this->moduleConfig->getConfigValue('pay_btn_text'),
+//                'showResponseMessage'       => false, // shows/hide the response popups
+//                'locale'                    => substr($locale, 0, 2),
+//                'autoOpenPM'                => (bool) $this->moduleConfig->getConfigValue('auto_expand_pms'),
+//                'logLevel'                  => $this->moduleConfig->getConfigValue('checkout_log_level'),
+//                'maskCvv'                   => true,
+//                'i18n'                      => $this->moduleConfig->getCheckoutTransl(),
+//                'blockCards'                => $blocked_cards,
+//                'theme'                     => $this->moduleConfig->getConfigValue('sdk_theme', 'checkout'),
+//                'apmWindowType'             => $this->moduleConfig->getConfigValue('apm_window_type', 'checkout'),
+//            ],
         ];
         
-        if ($isPaymentPlan) {
-            $sdk_data['nuveiCheckoutParams']['pmBlacklist'] = null;
-            $sdk_data['nuveiCheckoutParams']['pmWhitelist'] = ['cc_card'];
-        }
-        
-        if (in_array($save_pm, [true, 'always'])) {
-            $sdk_data['nuveiCheckoutParams']['userTokenId'] = $sdk_data['nuveiCheckoutParams']['email'];
-        }
+//        if ($isPaymentPlan) {
+//            $sdk_data['nuveiCheckoutParams']['pmBlacklist'] = null;
+//            $sdk_data['nuveiCheckoutParams']['pmWhitelist'] = ['cc_card'];
+//        }
+//        
+//        if (in_array($save_pm, [true, 'always'])) {
+//            $sdk_data['nuveiCheckoutParams']['userTokenId'] = $sdk_data['nuveiCheckoutParams']['email'];
+//        }
         
         return $sdk_data;
     }